Add cast to avoid signed vs unsigned comparison

+++ scilab/modules/abstractSyntaxTree/includes/runvisitor.hxx
@@ -185,7 +185,7 @@ namespace ast
 
 		types::InternalType* result_get(int _iPos)
 		{
-			if(_iPos >= _resultVect.size())
+			if(_iPos >= static_cast<int>(_resultVect.size()))
 			{
 				return NULL;
 			}
@@ -200,7 +200,7 @@ namespace ast
 		void result_set(int _iPos, const types::InternalType *gtVal)
 		{
 			m_bSingleResult = false;
-			if(_iPos < _resultVect.size())
+			if(_iPos <  static_cast<int>(_resultVect.size()))
 			{
 				if(_resultVect[_iPos] != NULL && _resultVect[_iPos]->isDeletable())
 				{
@@ -208,7 +208,7 @@ namespace ast
 				}
 			}
 
-			if(_iPos >= _resultVect.size())
+			if(_iPos >=  static_cast<int>(_resultVect.size()))
 			{
 				_resultVect.resize(_iPos + 1, NULL);
 			}
